fix: security hardening and code quality improvements from review
Security fixes:
- Remove setAccountAccess from MCP tools — account access is now
  user-only via the settings page, preventing rogue agents from
  weakening their own restrictions
- Add isFolderAccessible() check to all folderPath code paths
  (openFolder, listFolders, searchMessages, createFolder, renameFolder,
  deleteFolder, moveFolder, updateMessage moveTo, applyFilters) to
  prevent account access control bypass via direct folder URIs
- Add ./.​. guard on base64 attachment filenames

Lifecycle fixes:
- Store HTTP server on globalThis and call server.stop() in onShutdown
  so the port is released on extension reload
- Clear __tbMcpStartPromise in onShutdown to allow fresh restart
- Set start promise sentinel before awaiting to close double-start race

Code quality:
- Extract test/helpers.cjs with shared sendToBridge, sendMultipleToBridge,
  sanitizeJson, paginateResults, createValidator, connection file helpers
- Eliminate ~530 lines of duplicated code across 3 test files
- Normalize mixed tab/space indentation in api.js (89 lines)
