+++ test/auth.test.cjs
@@ -9,108 +9,20 @@
  * - Bridge behavior with and without auth
  */
 
-const { describe, it, before, after, beforeEach, afterEach } = require('node:test');
+const { describe, it, before, after } = require('node:test');
 const assert = require('node:assert/strict');
 const http = require('http');
 const fs = require('fs');
-const path = require('path');
-const os = require('os');
-const { spawn } = require('child_process');
 
-const net = require('net');
-
-const BRIDGE_PATH = path.resolve(__dirname, '..', 'mcp-bridge.cjs');
-const CONN_DIR = path.join(os.tmpdir(), 'thunderbird-mcp');
-const CONN_FILE = path.join(CONN_DIR, 'connection.json');
-const DEFAULT_PORT = 8765;
-
-/** Check if the default MCP port is already in use (e.g. real Thunderbird). */
-function isDefaultPortInUse() {
-  return new Promise((resolve) => {
-    const sock = net.createConnection({ port: DEFAULT_PORT, host: '127.0.0.1' });
-    sock.on('connect', () => { sock.destroy(); resolve(true); });
-    sock.on('error', () => resolve(false));
-  });
-}
-
-/**
- * Helper: send a JSON-RPC message to the bridge and get the response.
- */
-function sendToBridge(message, { timeout = 10000 } = {}) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(process.execPath, [BRIDGE_PATH], {
-      stdio: ['pipe', 'pipe', 'pipe'],
-    });
-
-    let stdout = '';
-    let stderr = '';
-    const timer = setTimeout(() => {
-      child.kill();
-      reject(new Error(`Bridge timed out. stdout: ${stdout}, stderr: ${stderr}`));
-    }, timeout);
-
-    child.stdout.on('data', (data) => {
-      stdout += data.toString();
-      const lines = stdout.split('\n').filter(l => l.trim());
-      if (lines.length > 0) {
-        clearTimeout(timer);
-        child.stdin.end();
-        try {
-          resolve(JSON.parse(lines[0]));
-        } catch (e) {
-          reject(new Error(`Failed to parse: ${lines[0]}`));
-        }
-      }
-    });
-
-    child.stderr.on('data', (data) => {
-      stderr += data.toString();
-    });
-
-    child.on('error', (err) => {
-      clearTimeout(timer);
-      reject(err);
-    });
-
-    child.on('exit', () => {
-      clearTimeout(timer);
-      if (!stdout.trim()) resolve(null);
-    });
-
-    child.stdin.write(JSON.stringify(message) + '\n');
-  });
-}
-
-/**
- * Write a test connection.json file.
- */
-function writeTestConnectionInfo(port, token) {
-  fs.mkdirSync(CONN_DIR, { recursive: true });
-  fs.writeFileSync(CONN_FILE, JSON.stringify({ port, token, pid: process.pid }), 'utf8');
-}
-
-/**
- * Back up and restore any existing connection file to avoid
- * interfering with a running Thunderbird instance.
- */
-let savedConnectionData = null;
-
-function backupConnectionFile() {
-  try {
-    savedConnectionData = fs.readFileSync(CONN_FILE, 'utf8');
-  } catch {
-    savedConnectionData = null;
-  }
-}
-
-function restoreConnectionFile() {
-  if (savedConnectionData !== null) {
-    fs.mkdirSync(CONN_DIR, { recursive: true });
-    fs.writeFileSync(CONN_FILE, savedConnectionData, 'utf8');
-  } else {
-    try { fs.unlinkSync(CONN_FILE); } catch { /* ignore */ }
-  }
-}
+const {
+  sendToBridge,
+  writeTestConnectionInfo,
+  backupConnectionFile,
+  restoreConnectionFile,
+  isDefaultPortInUse,
+  CONN_DIR,
+  CONN_FILE,
+} = require('./helpers.cjs');
 
 describe('Auth: connection info file', () => {
   before(() => backupConnectionFile());
@@ -122,7 +34,6 @@ describe('Auth: connection info file', () => {
     let receivedHeaders = null;
     let receivedPort = null;
 
-    // Start a mock server on the test port
     const server = http.createServer((req, res) => {
       receivedHeaders = req.headers;
       receivedPort = TEST_PORT;
@@ -140,7 +51,6 @@ describe('Auth: connection info file', () => {
     });
 
     try {
-      // Write connection info pointing to our mock server
       writeTestConnectionInfo(TEST_PORT, TEST_TOKEN);
 
       const response = await sendToBridge({
@@ -149,13 +59,8 @@ describe('Auth: connection info file', () => {
         method: 'tools/list'
       });
 
-      // Verify the bridge connected to our mock server (correct port)
       assert.equal(receivedPort, TEST_PORT);
-
-      // Verify the auth token was sent
       assert.equal(receivedHeaders['authorization'], `Bearer ${TEST_TOKEN}`);
-
-      // Verify we got a valid response
       assert.equal(response.id, 1);
     } finally {
       await new Promise((resolve) => server.close(resolve));
@@ -167,11 +72,8 @@ describe('Auth: connection info file', () => {
       return t.skip('Port 8765 in use (Thunderbird running), skipping fallback test');
     }
 
-    // Remove connection file
     try { fs.unlinkSync(CONN_FILE); } catch { /* ignore */ }
 
-    // The bridge should try port 8765 — we won't mock it, so we expect
-    // a connection error. We just verify it doesn't crash.
     const response = await sendToBridge({
       jsonrpc: '2.0',
       id: 2,
@@ -179,7 +81,6 @@ describe('Auth: connection info file', () => {
     });
 
     assert.equal(response.id, 2);
-    // Will be an error since no server on default port
     assert.ok(response.result || response.error);
   });
 });
@@ -192,7 +93,6 @@ describe('Auth: token verification', () => {
   before(async () => {
     backupConnectionFile();
 
-    // Mock server that checks auth like the extension does
     server = http.createServer((req, res) => {
       let authHeader = req.headers['authorization'] || '';
       if (authHeader !== `Bearer ${CORRECT_TOKEN}`) {
@@ -252,9 +152,7 @@ describe('Auth: token verification', () => {
       method: 'tools/list'
     });
 
-    // The bridge receives the 403 response body which has an error
-    // The mock returns a JSON-RPC error, so the bridge should parse it
-    assert.equal(response.id, null); // 403 response uses id: null
+    assert.equal(response.id, null);
     assert.ok(response.error);
     assert.match(response.error.message, /auth token/i);
   });
