+++ test/bridge.test.cjs
@@ -14,80 +14,14 @@ const { describe, it, before, after } = require('node:test');
 const assert = require('node:assert/strict');
 const http = require('http');
 const fs = require('fs');
-const os = require('os');
-const { spawn } = require('child_process');
 const path = require('path');
 
-const BRIDGE_PATH = path.resolve(__dirname, '..', 'mcp-bridge.cjs');
-
-/**
- * To test internal functions we need to extract them.
- * Since mcp-bridge.cjs has side effects (starts reading stdin), we test
- * the bridge through its stdio interface using child_process.
- */
-
-/**
- * Helper: send a JSON-RPC message to the bridge via stdin and read the response.
- * Returns a promise that resolves with the parsed JSON response.
- */
-function sendToBridge(message, { port, timeout = 5000 } = {}) {
-  return new Promise((resolve, reject) => {
-    const env = { ...process.env };
-    // We can't change the port in the bridge without modifying it,
-    // so for HTTP forwarding tests we'll use a mock server on port 8765
-    // or test only the locally-handled methods.
-    const child = spawn(process.execPath, [BRIDGE_PATH], {
-      stdio: ['pipe', 'pipe', 'pipe'],
-      env,
-    });
-
-    let stdout = '';
-    let stderr = '';
-    const timer = setTimeout(() => {
-      child.kill();
-      reject(new Error(`Bridge timed out after ${timeout}ms. stdout: ${stdout}, stderr: ${stderr}`));
-    }, timeout);
-
-    child.stdout.on('data', (data) => {
-      stdout += data.toString();
-      // Bridge outputs one JSON line per response
-      const lines = stdout.split('\n').filter(l => l.trim());
-      if (lines.length > 0) {
-        clearTimeout(timer);
-        child.stdin.end();
-        try {
-          resolve(JSON.parse(lines[0]));
-        } catch (e) {
-          reject(new Error(`Failed to parse bridge output: ${lines[0]}`));
-        }
-      }
-    });
-
-    child.stderr.on('data', (data) => {
-      stderr += data.toString();
-    });
-
-    child.on('error', (err) => {
-      clearTimeout(timer);
-      reject(err);
-    });
-
-    child.on('exit', (code) => {
-      clearTimeout(timer);
-      if (!stdout.trim()) {
-        // For notifications, the bridge returns nothing and exits cleanly
-        resolve(null);
-      }
-    });
-
-    child.stdin.write(JSON.stringify(message) + '\n');
-    // For notifications, close stdin immediately so the process exits
-    if (!Object.prototype.hasOwnProperty.call(message, 'id') ||
-        (typeof message.method === 'string' && message.method.startsWith('notifications/'))) {
-      child.stdin.end();
-    }
-  });
-}
+const {
+  sendToBridge,
+  sanitizeJson,
+  CONN_DIR,
+  CONN_FILE,
+} = require('./helpers.cjs');
 
 describe('Bridge: MCP lifecycle methods', () => {
   it('responds to initialize with protocol version and capabilities', async () => {
@@ -153,13 +87,11 @@ describe('Bridge: notifications', () => {
 describe('Bridge: HTTP forwarding', () => {
   let mockServer;
   let mockPort = 8765;
-  const connDir = path.join(os.tmpdir(), 'thunderbird-mcp');
-  const connFile = path.join(connDir, 'connection.json');
   let savedConnData = null;
 
   before(async () => {
     // Back up existing connection file
-    try { savedConnData = fs.readFileSync(connFile, 'utf8'); } catch { savedConnData = null; }
+    try { savedConnData = fs.readFileSync(CONN_FILE, 'utf8'); } catch { savedConnData = null; }
 
     // Start a mock server on port 8765 to simulate Thunderbird
     mockServer = http.createServer((req, res) => {
@@ -167,7 +99,6 @@ describe('Bridge: HTTP forwarding', () => {
       req.on('data', (chunk) => { body += chunk; });
       req.on('end', () => {
         const message = JSON.parse(body);
-        // Echo back a valid tools/list response
         if (message.method === 'tools/list') {
           res.writeHead(200, { 'Content-Type': 'application/json' });
           res.end(JSON.stringify({
@@ -196,7 +127,6 @@ describe('Bridge: HTTP forwarding', () => {
     await new Promise((resolve, reject) => {
       mockServer.on('error', (err) => {
         if (err.code === 'EADDRINUSE') {
-          // Port 8765 is in use (likely real Thunderbird) — skip these tests
           mockServer = null;
           resolve();
         } else {
@@ -204,9 +134,8 @@ describe('Bridge: HTTP forwarding', () => {
         }
       });
       mockServer.listen(mockPort, '127.0.0.1', () => {
-        // Write connection file so the bridge finds our mock server
-        fs.mkdirSync(connDir, { recursive: true });
-        fs.writeFileSync(connFile, JSON.stringify({ port: mockPort, token: 'mock-test-token' }), 'utf8');
+        fs.mkdirSync(CONN_DIR, { recursive: true });
+        fs.writeFileSync(CONN_FILE, JSON.stringify({ port: mockPort, token: 'mock-test-token' }), 'utf8');
         resolve();
       });
     });
@@ -216,12 +145,11 @@ describe('Bridge: HTTP forwarding', () => {
     if (mockServer) {
       await new Promise((resolve) => mockServer.close(resolve));
     }
-    // Restore original connection file
     if (savedConnData !== null) {
-      fs.mkdirSync(connDir, { recursive: true });
-      fs.writeFileSync(connFile, savedConnData, 'utf8');
+      fs.mkdirSync(CONN_DIR, { recursive: true });
+      fs.writeFileSync(CONN_FILE, savedConnData, 'utf8');
     } else {
-      try { fs.unlinkSync(connFile); } catch { /* ignore */ }
+      try { fs.unlinkSync(CONN_FILE); } catch { /* ignore */ }
     }
   });
 
@@ -253,56 +181,20 @@ describe('Bridge: HTTP forwarding', () => {
   });
 });
 
-describe('Bridge: sanitizeJson (via malformed response handling)', () => {
-  // We can't call sanitizeJson directly, but we can test the bridge's
-  // behavior when it receives responses with control characters.
-  // Instead, we'll test the sanitize logic inline.
-
+describe('Bridge: sanitizeJson', () => {
   it('removes control characters except newline, carriage return, tab', () => {
-    // Replicate the sanitizeJson logic for direct testing
-    function sanitizeJson(data) {
-      let sanitized = data.replace(/[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]/g, '');
-      sanitized = sanitized.replace(/(?<!\\)\r/g, '\\r');
-      sanitized = sanitized.replace(/(?<!\\)\n/g, '\\n');
-      sanitized = sanitized.replace(/(?<!\\)\t/g, '\\t');
-      return sanitized;
-    }
-
-    // Null byte and bell character should be removed
     assert.equal(sanitizeJson('hello\x00world\x07'), 'helloworld');
-
-    // Newline and tab should be escaped
     assert.equal(sanitizeJson('line1\nline2'), 'line1\\nline2');
     assert.equal(sanitizeJson('col1\tcol2'), 'col1\\tcol2');
-
-    // Already-escaped sequences should be preserved
     assert.equal(sanitizeJson('already\\nescaped'), 'already\\nescaped');
-
-    // Mixed control characters
     assert.equal(sanitizeJson('\x01\x02hello\nworld\x7f'), 'hello\\nworld');
   });
 
   it('handles empty string', () => {
-    function sanitizeJson(data) {
-      let sanitized = data.replace(/[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]/g, '');
-      sanitized = sanitized.replace(/(?<!\\)\r/g, '\\r');
-      sanitized = sanitized.replace(/(?<!\\)\n/g, '\\n');
-      sanitized = sanitized.replace(/(?<!\\)\t/g, '\\t');
-      return sanitized;
-    }
-
     assert.equal(sanitizeJson(''), '');
   });
 
   it('preserves valid JSON with unicode', () => {
-    function sanitizeJson(data) {
-      let sanitized = data.replace(/[\x00-\x08\x0b\x0c\x0e-\x1f\x7f]/g, '');
-      sanitized = sanitized.replace(/(?<!\\)\r/g, '\\r');
-      sanitized = sanitized.replace(/(?<!\\)\n/g, '\\n');
-      sanitized = sanitized.replace(/(?<!\\)\t/g, '\\t');
-      return sanitized;
-    }
-
     const input = '{"emoji": "\\ud83d\\ude00", "text": "hello"}';
     assert.equal(sanitizeJson(input), input);
   });
